Refactor pasapi applications
